requirements.txt
pykeepass==4.1.0.post1
python-dotenv==1.0.1
colorama==0.4.6
pytest==7.4.4
pytest-html==4.1.1
pytest-cov==4.1.0
